feat: P0 security hardening + SDK/CLI signing stack
btcpc-node:
- Remove /api/entry raw endpoint (MineReward injection vector)
- Explicit allowlist in tx.rs — Mine/MineReward/GenesisAlloc/EpochFinalize rejected externally
- Keyless account spending blocked (require_key guard on Transfer/Stake/Unstake)
- Write lock on Chain (parking_lot::Mutex) — serialises nonce-check → debit → bump
- EpochSeal epoch drift bounds: reject seals > current+3 or < current-10
- Atomic contract storage commit — storage writes happen only after all transfers succeed
- Deploy nonce: contract_id derived from actual nonce, deployer nonce checked and bumped
- Block gossip validation: reject blocks >+10 epochs ahead, GenesisAlloc/EpochFinalize entries, MineReward without Mine

btcpc-sdk:
- all_balances, account, epoch_meta endpoints
- get_nonce / next_nonce helpers
- KeyPair with generate, from_hex, from_file, sign_entry_json, sign_bytes
- Bsp20Client and Bsp721Client typed wrappers with auto-nonce + signing
- Amount fields: u64 dreams throughout (no f64 precision loss)

btcpc-cli:
- transfer/stake/unstake: auto-fetch nonce, sign canonical JSON, submit with signature
- btcpc key generate/show/register subcommands

+++ rust/btcpc-cli/src/contract.rs
@@ -65,7 +65,7 @@ pub fn cmd_contract_call(
     method: &str,
     signer: &str,
     args: Option<&str>,
-    deposit: Option<f64>,
+    deposit: Option<u64>,
 ) -> Result<()> {
     let api = ApiClient::new();
     let mut body = json!({
